feat: marge home page with introduction pafe
