+++ src/components/base/home/MainBanner.tsx
